Good that at least the server is on warnings as errors, fixing size_t problem

+++ Server/Source/SendThread.cpp
@@ -22,7 +22,7 @@ void SendThread::run()
 
 		// Now serialize the buffer and create the datagram to send back to the client
 		JammerNetzAudioData dataForClient(nextBlock.audioBlock);
-		int bytesWritten;
+		size_t bytesWritten;
 		dataForClient.serialize(writebuffer, bytesWritten);
 
 		if (fecData_.find(nextBlock.targetAddress) == fecData_.end()) {
